#383 - Move joomla specific logic to ComKoowaUserTrait
The `setUser()`method has been removed and the re-implemented as part of
the `setData()`method.
